Refactor code regarding lightcards to conform to the new UI Software architecture (#4083)
