+++ ui/main/src/app/business/services/lightcards/filter.service.ts
