+++ ui/main/src/app/business/services/lightcards/search-service.ts
