+++ ui/main/src/app/modules/activityarea/activityarea.component.ts
@@ -17,7 +17,7 @@ import {SettingsService} from 'app/business/services/settings.service';
 import {EntitiesService} from 'app/business/services/entities.service';
 import {Utilities} from '../../business/common/utilities';
 import {GroupsService} from 'app/business/services/groups.service';
-import {LightCardsStoreService} from '@ofServices/lightcards/lightcards-store.service';
+import {LightCardsStoreService} from 'app/business/services/lightcards/lightcards-store.service';
 import {ApplicationEventsService} from 'app/business/services/application-events.service';
 
 @Component({
