+++ ui/main/src/app/modules/feed/components/card-list/filters/feed-filter/feed-filter.component.ts
@@ -18,8 +18,8 @@ import {UserPreferencesService} from '@ofServices/user-preference.service';
 import {DateTimeNgb} from '@ofModel/datetime-ngb.model';
 import moment from 'moment';
 import {MessageLevel} from '@ofModel/message.model';
-import {FilterService} from '@ofServices/lightcards/filter.service';
-import {LightCardsFeedFilterService} from '@ofServices/lightcards/lightcards-feed-filter.service';
+import {FilterService} from 'app/business/services/lightcards/filter.service';
+import {LightCardsFeedFilterService} from 'app/business/services/lightcards/lightcards-feed-filter.service';
 import {Utilities} from 'app/business/common/utilities';
 import {AlertMessageService} from 'app/business/services/alert-message.service';
