+++ ui/main/src/app/modules/feed/components/time-line/init-chart/init-chart.component.spec.ts
@@ -30,13 +30,13 @@ import {HttpClient, HttpHandler} from '@angular/common/http';
 import {AppService} from '@ofServices/app.service';
 import {BusinessconfigI18nLoaderFactory} from '@tests/helpers';
 import {LightCardsServiceMock} from '@tests/mocks/lightcards.service.mock';
-import {LightCardsFeedFilterService} from '@ofServices/lightcards/lightcards-feed-filter.service';
+import {LightCardsFeedFilterService} from 'app/business/services/lightcards/lightcards-feed-filter.service';
 import {OpfabLoggerService} from 'app/business/services/logs/opfab-logger.service';
 import {RemoteLoggerService} from 'app/business/services/logs/remote-logger.service';
-import {RemoteLoggerServiceMock} from '@tests/mocks/remote-logger.service.mock';
 import {DateTimeFormatterService} from 'app/business/services/date-time-formatter.service';
 import {ConfigServer} from 'app/business/server/config.server';
 import {ConfigServerMock} from '@tests/mocks/configServer.mock';
+import {RemoteLoggerServiceMockAngular} from '@tests/mocks/remote-logger.service-angular.mock';
 
 describe('InitChartComponent', () => {
     let component: InitChartComponent;
@@ -83,7 +83,7 @@ describe('InitChartComponent', () => {
                 {provide: GlobalStyleService, useClass: GlobalStyleService},
                 {provide: LightCardsFeedFilterService, useClass: LightCardsServiceMock},
                 {provide: OpfabLoggerService, useClass: OpfabLoggerService},
-                {provide: RemoteLoggerService, useClass: RemoteLoggerServiceMock}
+                {provide: RemoteLoggerService, useClass: RemoteLoggerServiceMockAngular}
             ],
             schemas: [NO_ERRORS_SCHEMA]
         }).compileComponents();
