+++ ui/main/src/app/modules/monitoring/monitoring.component.ts
@@ -17,7 +17,7 @@ import {MonitoringFiltersComponent} from './components/monitoring-filters/monito
 import {Process, TypeOfStateEnum} from '@ofModel/processes.model';
 import {ProcessesService} from 'app/business/services/processes.service';
 import {Filter} from '@ofModel/feed-filter.model';
-import {LightCardsStoreService} from '@ofServices/lightcards/lightcards-store.service';
+import {LightCardsStoreService} from 'app/business/services/lightcards/lightcards-store.service';
 import {EntitiesService} from 'app/business/services/entities.service';
 
 @Component({
