+++ ui/main/src/app/services/sound-notification.service.ts
@@ -12,8 +12,8 @@ import {Injectable, OnDestroy} from '@angular/core';
 import {PlatformLocation} from '@angular/common';
 import {LightCard, Severity} from '@ofModel/light-card.model';
 import {Notification} from '@ofModel/external-devices.model';
-import {LightCardsFeedFilterService} from './lightcards/lightcards-feed-filter.service';
-import {LightCardsStoreService} from './lightcards/lightcards-store.service';
+import {LightCardsFeedFilterService} from '../business/services/lightcards/lightcards-feed-filter.service';
+import {LightCardsStoreService} from '../business/services/lightcards/lightcards-store.service';
 import {EMPTY, iif, merge, of, Subject, timer} from 'rxjs';
 import {filter, map, switchMap, takeUntil} from 'rxjs/operators';
 import {ExternalDevicesService} from 'app/business/services/external-devices.service';
